CP-C: maps routes, System map, and the canvas monolith decomposed
One graph, many maps becomes navigable (docs/spec/maps.md):

- /project/[id]/maps — the maps index: built-ins + custom maps from
  project.metadata.maps, subgraph counts per card, create/edit/delete
  custom maps in a dialog (humans get the dialog; agents write the same
  JSON — the seed's recording-loop map renders untouched)
- /project/[id]/maps/[mapId] — renderer shell: journey → JourneyMap,
  system → SystemMap, unknown ids and unrenderable kinds get inline cards
- System map: the model-centered reading — all four species as cards,
  every surviving cross-layer edge drawn, ELK-partitioned into species
  tiers (views / APIs / data models; spike-verified that
  elk.separateConnectedComponents=false pins edge-less nodes into their
  tier); connect-to-create and edge deletion reuse the Journey dialogs
- the 1,587-line canvas page decomposes: pure buildJourneyGraph +
  computeComposeClosure + computeViewApiRelations (lib/utils/
  journey-graph.ts), shared visual-id/species helpers (graph-build.ts),
  useElkLayout hook, RawBundleSheet extraction, and the JourneyMap
  component (scoped journeys honor definition.root_node_id); /canvas and
  /project/[id] redirect to /maps/journey so old links keep working
- golden parity test pins buildJourneyGraph over the Pebbles seed to the
  live-verified counts (22 collapsed / 25 expanded / 8 top-level flows /
  orphan fallback) — tests/app/journey-graph.test.js + CI step
- viewport fitting hardened: ReactFlow instance minZoom lowered so
  whole-product maps frame in full; SystemMap re-frames when the first
  ELK layout lands (derived fit signal)
- sidebar Maps group: All maps / Journey / System / custom maps;
  dead navigation code deleted (useGraphNavigation, Breadcrumb,
  PlatformDots); transient test-build dirs added to eslint ignores
- docs ride along: architecture routes + component map, graph-model
  sources, conventions, vision CP table marks CP-C shipped, maps spec
  banner says implemented and its edge-keeping wording now matches the
  induced-subgraph semantics the tests pin
