+++ .github/workflows/ci.yml
@@ -63,6 +63,9 @@ jobs:
       - name: Delivery board grouping tests
         run: npm run test:delivery
 
+      - name: Journey graph golden parity tests
+        run: npm run test:journey-graph
+
       - name: Journal write-primitive (emit) tests
         run: npm run test:emit
