--- components/layout/Breadcrumb.tsx
@@ -1,46 +1,0 @@
-import {
-  Breadcrumb as BreadcrumbRoot,
-  BreadcrumbItem,
-  BreadcrumbLink,
-  BreadcrumbList,
-  BreadcrumbPage,
-  BreadcrumbSeparator,
-} from "@/components/ui/breadcrumb";
-
-interface BreadcrumbSegment {
-  label: string;
-  href?: string;
-  onClick?: () => void;
-}
-
-interface BreadcrumbProps {
-  segments: BreadcrumbSegment[];
-}
-
-export function Breadcrumb({ segments }: BreadcrumbProps) {
-  return (
-    <BreadcrumbRoot>
-      <BreadcrumbList>
-        {segments.map((segment, index) => (
-          <span key={segment.label} className="flex items-center gap-1.5">
-            <BreadcrumbItem>
-              {segment.href ? (
-                <BreadcrumbLink href={segment.href}>{segment.label}</BreadcrumbLink>
-              ) : segment.onClick ? (
-                <button
-                  onClick={segment.onClick}
-                  className="hover:text-foreground transition-colors"
-                >
-                  {segment.label}
-                </button>
-              ) : (
-                <BreadcrumbPage>{segment.label}</BreadcrumbPage>
-              )}
-            </BreadcrumbItem>
-            {index < segments.length - 1 && <BreadcrumbSeparator />}
-          </span>
-        ))}
-      </BreadcrumbList>
-    </BreadcrumbRoot>
-  );
-}
