+++ components/layout/ProjectSwitcher.tsx
@@ -23,7 +23,7 @@ import { useProjects } from "@/lib/hooks/useProjects";
 interface ProjectSwitcherProps {
   currentProjectId: string;
   currentProjectTitle?: string;
-  currentView: "canvas" | "library" | "delivery" | "changelog";
+  currentView: "maps" | "library" | "delivery" | "changelog";
   currentQueryString?: string;
 }
