+++ package.json
@@ -21,6 +21,7 @@
     "test:journal": "node tests/schema/journal.test.js",
     "test:maps": "node tests/schema/maps.test.js",
     "test:delivery": "node tests/app/delivery.test.js",
+    "test:journey-graph": "node tests/app/journey-graph.test.js",
     "test:journal-projections": "node tests/data/journal-projections.test.js",
     "test:emit": "node tests/schema/emit.test.js",
     "test:emit-events": "node tests/data/emit-events.test.js",
