+++ tests/app/load-journey-graph.js
@@ -1,0 +1,65 @@
+/**
+ * Loads lib/utils/journey-graph.ts (the Journey map's pure graph construction)
+ * into Node without a bundler — the load-delivery.js technique over its small
+ * runtime graph: graph-build.ts and platform-status.ts plus the config const
+ * arrays. All `@xyflow/react` and `@arkaik/schema` imports in this graph are
+ * type-only (erased).
+ */
+
+const fs = require("fs");
+const path = require("path");
+const ts = require("typescript");
+
+const ROOT = path.join(__dirname, "..", "..");
+const BUILD_DIR = path.join(__dirname, ".test-build-journey-graph");
+
+// Source path (repo-relative) → build output basename.
+const MODULES = [
+  ["lib/config/platforms.ts", "config-platforms"],
+  ["lib/config/statuses.ts", "config-statuses"],
+  ["lib/utils/platform-status.ts", "platform-status"],
+  ["lib/utils/graph-build.ts", "graph-build"],
+  ["lib/utils/journey-graph.ts", "journey-graph"],
+];
+
+// `@/lib/...` specifier → build output basename.
+const SPECIFIER_MAP = {
+  "@/lib/config/platforms": "./config-platforms",
+  "@/lib/config/species": "./config-species", // type-only in this graph
+  "@/lib/config/statuses": "./config-statuses",
+  "@/lib/config/edge-types": "./config-edge-types", // type-only in this graph
+  "@/lib/data/types": "./types", // type-only in this graph
+  "@/lib/utils/platform-status": "./platform-status",
+  "@/lib/utils/graph-build": "./graph-build",
+};
+
+function loadJourneyGraph() {
+  fs.rmSync(BUILD_DIR, { recursive: true, force: true });
+  fs.mkdirSync(BUILD_DIR, { recursive: true });
+  fs.writeFileSync(path.join(BUILD_DIR, "package.json"), JSON.stringify({ type: "commonjs" }));
+
+  for (const [srcRel, outName] of MODULES) {
+    const source = fs.readFileSync(path.join(ROOT, srcRel), "utf8");
+    const { outputText } = ts.transpileModule(source, {
+      fileName: path.basename(srcRel),
+      compilerOptions: {
+        module: ts.ModuleKind.CommonJS,
+        target: ts.ScriptTarget.ES2020,
+        esModuleInterop: true,
+      },
+    });
+
+    let rewritten = outputText;
+    for (const [specifier, target] of Object.entries(SPECIFIER_MAP)) {
+      rewritten = rewritten.split(`require("${specifier}")`).join(`require("${target}")`);
+    }
+    fs.writeFileSync(path.join(BUILD_DIR, `${outName}.js`), rewritten);
+  }
+
+  for (const [, outName] of MODULES) {
+    delete require.cache[path.join(BUILD_DIR, `${outName}.js`)];
+  }
+  return require(path.join(BUILD_DIR, "journey-graph.js"));
+}
+
+module.exports = { loadJourneyGraph, BUILD_DIR };
